Big rename: FileTree -> BlobTree;  File -> Blob
These new names reflect the reality that we're not expecting a full
file-like interface for these things. The intent is to support more
"plain data" like things than the complex OS-dependent metadata that
real filesystems support.

Also FileTrees.jl has recently taken the name `FileTree` and it's too
bad to have a direct clash of names for packages which are rather
similar and could be used together.