WIP: Data Layers
Data layers allow data of different formats to be mapped into a program
through a decoder and presented with a uniform API such that the main
program logic can avoid dealing with data format decoding. Instead, the
data format can be defined in the Data.toml.

A challenge here is dealing with world age issues which come up from
dynamically `require`ing Julia packages. For now, we include a bit of
judicious Base.invokelatest to make things "just work" in the REPL, but
also warn the user that they should add a top-level import.

Here's some examples of data which comes encoded in many forms, but we'd
like to treat through the same API:

Tabular data:

* csv
* gzip.csv
* tsv
* arrow
* parquet
* ...

Byte streams:

* raw
* gzip
* xz
* zstd
* ...

Images

* png
* jpeg
* tiff
* ...

Data trees

* directories
* zip
* hdf5