Flexible string-based data addressing with URL-like "dataspec" (#33)
This parses the argument to dataset() with a URL-like scheme, enabling
more flexible naming of resources within a given dataset. Like URLs, we
use the & and # as separators for query and fragment sections.

Like URLs, this provides for a compact string representation which can
be easily communicated between users of a shared dataset to address
subresources without the need to create an entirely new and largely
duplicated dataset configuration.

As in URLs, the query section is to be used during resource resolution,
for example a version number for the dataset can be specified.  We
choose the standard & and = separators for key value pairs:

    name/of/dataset?version=v1

As in URLs, the fragment section refers to a subresource and its
interpretation depends on the type of the dataset. (In analogy, the
meaning of the fragment of a URL depends on the mime type received from
the server.) For example, a subtree of a BlobTree can be addressed using

tree_name#path/to/subtree

The use of fragment for indexing a BlobTree is implemented here as one
handy use case of this syntax.