Big BlobTree API refactor including newfile() / newdir()
Refactor newfile() / newdir() into primarily in-place APIs. A
temporary directory can still be created with newdir().
Before this refactor the API looked slightly nicer, but we would never
be able to reach the efficiency of the native file APIs. In
particular, we might end up moving data across devices as a last step
in constructing a directory tree — this seemed bad!

Various changes to BlobTree API to make it more Dict-like and less
reliant on overloading filesystem-like functions.

* Further document the BlobTree API
* Allow path strings as keys in more places in BlobTree
* Deprecations of some filesystem-like functions