Add DataSets.from_path(); allow `-` in dataset names (#40)
The from_path function can be used to create a standalone DataSet from
data on the local filesystem, inferring the type as file or directory.

Also allow `-` in dataset names: this should be harmless enough in terms
of our URL-like dataspec syntax, and also in the data REPL where
shell-like splitting treats `-` as part of words. This is particularly
useful on JuliaHub where we allow user names to contain `-`.