Deprecate `@__DIR__` templating for use in Data.toml
Templating paths with `@__DIR__` was always a hack because it was
unaware of the structure of the TOML file and didn't generalize well.

Replace this with querying of a `DataSet`s parent project for its
absolute path (or rather, by asking the project to join its root path to
the relative path contained in the Data.toml).

This allows DataSet instances to be safely saved back to their parent
projects.